RavenDB-18232 - fixing PR comments#1
