+++ src/Raven.Server/Documents/Sharding/Handlers/Processors/OngoingTasks/ShardedOngoingTasksHandlerProcessorForGetOngoingTaskInfo.cs
@@ -1,0 +1,17 @@
+﻿using System.Threading.Tasks;
+using JetBrains.Annotations;
+
+namespace Raven.Server.Documents.Sharding.Handlers.Processors.OngoingTasks
+{
+    internal class ShardedOngoingTasksHandlerProcessorForGetOngoingTaskInfo : ShardedOngoingTasksHandlerProcessorForGetOngoingTasksInfo
+    {
+        public ShardedOngoingTasksHandlerProcessorForGetOngoingTaskInfo([NotNull] ShardedDatabaseRequestHandler requestHandler) : base(requestHandler)
+        {
+        }
+
+        public override async ValueTask ExecuteAsync()
+        {
+            await GetOngoingTaskInfoInternalAsync();
+        }
+    }
+}
