+++ src/Raven.Server/Web/System/Processors/OngoingTasks/OngoingTasksHandlerProcessorForGetOngoingTaskInfo.cs
@@ -1,0 +1,18 @@
+﻿using System.Threading.Tasks;
+using JetBrains.Annotations;
+using Raven.Server.Documents;
+
+namespace Raven.Server.Web.System.Processors.OngoingTasks
+{
+    internal class OngoingTasksHandlerProcessorForGetOngoingTaskInfo : OngoingTasksHandlerProcessorForGetOngoingTasksInfo
+    {
+        public OngoingTasksHandlerProcessorForGetOngoingTaskInfo([NotNull] DatabaseRequestHandler requestHandler) : base(requestHandler)
+        {
+        }
+
+        public override async ValueTask ExecuteAsync()
+        {
+            await GetOngoingTaskInfoInternalAsync();
+        }
+    }
+}
